pyproject.toml
[tool.poetry]
name = "tkfeather"
version = "1.0.1"
description = "Feather icons for tkinter"
authors = ["John Riggles <jriggles@icloud.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/JRiggles/tkfeather"
keywords = ["tkinter", "icons", "Feather"]

[tool.poetry.dependencies]
python = "^3.12"
pillow = ">=9.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
